backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-multipart==0.0.20
pydantic==2.10.4
email-validator==2.2.0
requests==2.32.3
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-dotenv==1.0.1
PyJWT
